+++ apps/fyler/src/handlers/doc_to_pdf.erl
@@ -7,7 +7,7 @@
 
 -export([run/1,run/2, category/0]).
 
--define(COMMAND(In), "unoconv -f pdf " ++ In).
+-define(COMMAND(In), io_lib:format("unoconv -f pdf ~s",[In])).
 
 category() ->
   document.
@@ -18,14 +18,14 @@ run(#file{tmp_path = Path, name = Name, dir = Dir},_Opts) ->
   Start = ulitos:timestamp(),
   ?D({"command",?COMMAND(Path)}),
   Data = os:cmd(?COMMAND(Path)),
-  PDF = Dir ++ "/" ++ Name ++ ".pdf",
+  PDF = filename:join(Dir,Name ++ ".pdf"),
   case  filelib:is_file(PDF) of
     true -> case pdf_thumb:run(#file{tmp_path = PDF, name = Name, dir = Dir}) of
               {ok,#job_stats{result_path = Thumb}} ->
                     {ok,#job_stats{time_spent = ulitos:timestamp() - Start, result_path = [list_to_binary(Name ++ ".pdf")|Thumb]}};
               Else -> Else
             end;
-    _ -> {error, Data}
+    _ -> {error, {unoconv_failed,PDF,Data}}
   end.