+++ apps/fyler/src/handlers/pdf_to_pages.erl
@@ -17,7 +17,7 @@ run(File) -> run(File, []).
 
 run(#file{tmp_path = Path, name = Name, dir = Dir}, _Opts) ->
   Start = ulitos:timestamp(),
-  PagesDir = Dir ++ "/pages",
+  PagesDir = filename:join(Dir,"pages"),
   ok = file:make_dir(PagesDir),
   ?D({"command", ?COMMAND(Path, PagesDir ++ "/")}),
   Data = os:cmd(?COMMAND(Path, PagesDir ++ "/")),
@@ -27,7 +27,7 @@ run(#file{tmp_path = Path, name = Name, dir = Dir}, _Opts) ->
     List ->
       List2 = case make_progressive_jpeg(PagesDir, List) of
                 {ok, List_} ->
-                  [file:delete(PagesDir ++ "/" ++ Old) || Old <- List],
+                  [file:delete(filename:join(PagesDir,Old)) || Old <- List],
                   List_;
                 false -> List
               end,
@@ -39,7 +39,7 @@ run(#file{tmp_path = Path, name = Name, dir = Dir}, _Opts) ->
           {pages, [list_to_binary(T) || T <- List2]}
         ]
       }),
-      JSONFile = Dir ++ "/" ++ Name ++ ".pages.json",
+      JSONFile = filename:join(Dir,Name ++ ".pages.json"),
       {ok, F} = file:open(JSONFile, [write]),
       file:write(F, JSON),
       file:close(F),
@@ -49,7 +49,7 @@ run(#file{tmp_path = Path, name = Name, dir = Dir}, _Opts) ->
 
 
 make_progressive_jpeg(Dir, List) ->
-  [os:cmd(?COMMAND2(Dir ++ "/" ++ F, Dir ++ "/pr_" ++ F)) || F <- List],
+  [os:cmd(?COMMAND2(filename:join(Dir,F), filename:join(Dir,"pr_" ++ F))) || F <- List],
   case filelib:wildcard("pr_*.jpg", Dir) of
     [] -> ?D({error_converting_to_progressive, Dir}), false;
     List2 -> {ok, List2}