+++ apps/fyler/src/tests/conversions_tests.erl
@@ -1,0 +1,302 @@
+-module(conversions_tests).
+-include("../fyler.hrl").
+-include("../../include/log.hrl").
+-include_lib("kernel/include/file.hrl").
+-include_lib("eunit/include/eunit.hrl").
+
+-define(PATH(File), filename:join([code:priv_dir(fyler),"..","src","tests","support", File])).
+
+-define(setup(F), {setup, fun setup_/0, fun cleanup_/1, F}).
+
+-define(TIMEOUT, 60).
+
+file_size(Path) ->
+  case file:read_file_info(Path) of
+    {ok,#file_info{size = Size}} -> Size;
+    _ -> 0
+  end.
+
+video_probe_flv_speex_test() ->
+  ?assertMatch(#video_info{audio_codec="libspeex", video_codec="h264"}, video_probe:info(?PATH("stream_1.flv"))).
+
+video_probe_screen_share_test() ->    
+  ?assertMatch(#video_info{video_codec="flashsv2", pixel_format="bgr24"}, video_probe:info(?PATH("stream_2.flv"))).
+
+video_probe_mov_test() ->
+  ?assertMatch(#video_info{video_codec="mjpeg"}, video_probe:info(?PATH("v1.MOV"))).
+
+video_probe_aac_h264_test() ->
+  ?assertMatch(#video_info{video_codec="h264", audio_codec="aac"}, video_probe:info(?PATH("v2.mp4"))).
+
+video_probe_avi_test() ->
+  ?assertMatch(#video_info{video_codec="mpeg4"}, video_probe:info(?PATH("v3.avi"))).
+
+delete_files([]) ->
+  ok;
+
+delete_files([File|Files]) ->
+  file:delete(?PATH(File)),
+  delete_files(Files).
+
+setup_() ->
+  lager:start(),
+  file:make_dir(?PATH("tmp")).
+
+cleanup_(_) ->
+  ulitos_file:recursively_del_dir(?PATH("tmp")),
+  ulitos_file:recursively_del_dir(?PATH("thumbs")),
+  ulitos_file:recursively_del_dir(?PATH("swfs")),
+  file:delete(?PATH("10.pdf")),
+  file:delete(?PATH("13.pdf")),
+  file:delete(?PATH("2.pdf")),
+  file:delete(?PATH("v1.mp4")),
+  file:delete(?PATH("v2_converted.mp4")),
+  file:delete(?PATH("v3.mp4")),
+  file:delete(?PATH("8.mp3")),
+  file:delete(?PATH("9_converted.mp3")),
+  file:delete(?PATH("3.pdf")),
+  file:delete(?PATH("6.pdf")),
+  file:delete(?PATH("7_split.pdf")),
+  case filelib:wildcard("*.ts",?PATH("")) of
+    Files when is_list(Files) -> delete_files(Files);
+    _ -> false
+  end,
+  case filelib:wildcard("*.json",?PATH("")) of
+    Files2 when is_list(Files2) -> delete_files(Files2);
+    _ -> false
+  end,
+  case filelib:wildcard("*.m3u8",?PATH("")) of
+    Files3 when is_list(Files3) -> delete_files(Files3);
+    _ -> false
+  end,
+  case filelib:wildcard("*.png",?PATH("")) of
+    Files4 when is_list(Files4) -> delete_files(Files4);
+    _ -> false
+  end,
+  application:stop(lager).
+
+
+
+unoconv_test_() ->
+  ?setup([
+    {timeout, ?TIMEOUT, [doc_to_pdf_t_()]},
+    {timeout, ?TIMEOUT, [doc_to_pdf_2_t_()]},
+    {timeout, ?TIMEOUT, [xls_to_pdf_t_()]},
+    {timeout, ?TIMEOUT, [xls_to_pdf_2_t_()]},
+    {timeout, ?TIMEOUT, [ppt_to_pdf_t_()]}
+   ]
+  ).
+
+swftools_test_() ->
+  ?setup(
+    [
+      {timeout, ?TIMEOUT, [doc_to_pdf_swf_t_()]},
+      {timeout, ?TIMEOUT, [doc_to_pdf_thumbs_t_()]},
+      {timeout, ?TIMEOUT, [pdf_to_swf_t_()]},
+      {timeout, ?TIMEOUT, [pdf_to_swf_thumbs_t_()]}
+    ]
+  ).
+
+pdftk_test_()->
+  ?setup(
+    [ 
+      {timeout, ?TIMEOUT, [pdf_split_t_()]}  
+    ]
+  ).
+
+video_mp4_test_() ->
+  ?setup(
+    [
+      {timeout, ?TIMEOUT, [mov_to_mp4_t_()]},
+      {timeout, ?TIMEOUT, [avi_to_mp4_t_()]},
+      {timeout, ?TIMEOUT, [mp4_to_mp4_t_()]}
+    ]
+  ).
+
+video_hls_test_() ->
+  ?setup(
+    [
+      {timeout, ?TIMEOUT, [mov_to_hls_t_()]},
+      {timeout, ?TIMEOUT, [avi_to_hls_t_()]},
+      {timeout, ?TIMEOUT, [mp4_to_hls_t_()]}
+    ]
+  ).
+
+video_recording_test_() ->
+  ?setup(
+    [
+      {timeout, ?TIMEOUT, [av_rec_to_hls_t_()]},
+      {timeout, ?TIMEOUT, [screen_rec_to_hls_t_()]}
+    ]
+  ).
+
+audio_to_mp3_test_() ->
+  ?setup(
+    [
+      {timeout, ?TIMEOUT, [wav_to_mp3_t_()]},
+      {timeout, ?TIMEOUT, [mp3_to_mp3_t_()]}
+    ]
+  ).
+
+doc_to_pdf_t_() ->
+  fun() ->
+    Res = doc_to_pdf:run(#file{tmp_path = ?PATH("2.docx"), name = "2", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.     
+
+doc_to_pdf_2_t_() ->
+  fun() ->
+    Res = doc_to_pdf:run(#file{tmp_path = ?PATH("3.doc"), name = "3", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.     
+
+xls_to_pdf_t_() ->
+  fun() ->
+    Res = doc_to_pdf:run(#file{tmp_path = ?PATH("6.xlsx"), name = "6", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.    
+
+xls_to_pdf_2_t_() ->
+  fun() ->
+    Res = doc_to_pdf:run(#file{tmp_path = ?PATH("10.xls"), name = "10", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.  
+
+ppt_to_pdf_t_() ->
+  fun() ->
+    Res = doc_to_pdf:run(#file{tmp_path = ?PATH("13.pptx"), name = "13", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.     
+
+doc_to_pdf_swf_t_() ->
+  fun() ->
+    Res = doc_to_pdf_swf:run(#file{tmp_path = ?PATH("2.docx"), name = "2", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(3, length(Stat#job_stats.result_path))
+  end.  
+
+doc_to_pdf_thumbs_t_() ->
+  fun() ->
+    Res = doc_to_pdf_thumbs:run(#file{tmp_path = ?PATH("2.docx"), name = "2", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end.  
+
+pdf_split_t_() ->
+  fun() ->
+    OldSize = file_size(?PATH("7.pdf")),
+    Res = split_pdf:run(#file{tmp_path = ?PATH("7.pdf"), name = "7", dir = ?PATH("")},[{split,<<"10-20 25-end">>}]),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path)),
+    NewSize = file_size(?PATH("7_split.pdf")),
+    ?assert(OldSize>NewSize)
+  end.
+
+pdf_to_swf_t_() ->
+  fun() ->
+    Res = pdf_to_swf:run(#file{tmp_path = ?PATH("7.pdf"), name = "7", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end. 
+
+pdf_to_swf_thumbs_t_() ->
+  fun() ->
+    Res = pdf_to_swf_thumbs:run(#file{tmp_path = ?PATH("7.pdf"), name = "7", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(2, length(Stat#job_stats.result_path))
+  end. 
+
+mov_to_mp4_t_() ->
+  fun() ->
+    Res = video_to_mp4:run(#file{tmp_path = ?PATH("v1.MOV"), name = "v1", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(3, length(Stat#job_stats.result_path))
+  end.
+
+avi_to_mp4_t_() ->
+  fun() ->
+    Res = video_to_mp4:run(#file{tmp_path = ?PATH("v3.avi"), name = "v3", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(3, length(Stat#job_stats.result_path))
+  end.
+
+mp4_to_mp4_t_() ->
+  fun() ->
+    Res = video_to_mp4:run(#file{tmp_path = ?PATH("v2.mp4"), name = "v2", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(3, length(Stat#job_stats.result_path))
+  end.
+
+mov_to_hls_t_() ->
+  fun() ->
+    Res = video_to_hls:run(#file{tmp_path = ?PATH("v1.MOV"), name = "v1", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+mp4_to_hls_t_() ->
+  fun() ->
+    Res = video_to_hls:run(#file{tmp_path = ?PATH("v2.mp4"), name = "v2", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+avi_to_hls_t_() ->
+  fun() ->
+    Res = video_to_hls:run(#file{tmp_path = ?PATH("v3.avi"), name = "v3", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+av_rec_to_hls_t_() ->
+  fun() ->
+    Res = recording_to_hls:run(#file{tmp_path = ?PATH("stream_1.flv"), name = "stream_1", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+screen_rec_to_hls_t_() ->
+  fun() ->
+    Res = recording_to_hls:run(#file{tmp_path = ?PATH("stream_2.flv"), name = "stream_1", dir = ?PATH("")},[{stream_type,<<"share">>}]),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+wav_to_mp3_t_() ->
+  fun() ->
+    Res = audio_to_mp3:run(#file{tmp_path = ?PATH("8.wav"), name = "8", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.
+
+mp3_to_mp3_t_() ->
+  fun() ->
+    Res = audio_to_mp3:run(#file{tmp_path = ?PATH("9.mp3"), name = "9", dir = ?PATH("")}),
+    ?assertMatch({ok,#job_stats{}}, Res),
+    {_, Stat} = Res,
+    ?assertEqual(1, length(Stat#job_stats.result_path))
+  end.